requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
prometheus-client==0.20.0
aiohttp==3.9.1
